+++ CHANGELOG.md
@@ -12,6 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [5.0.0]- 2024.09.04
+
+### Added
+
+- Added the `createHash`, `createPasswordHash` and `isPasswordHashValid` methods to the `App` utility class.
+
+### Changed
+
+- Changed all database migrations to use the `bones.migrations` filter.
+- Updated GitHub issue templates.
+- Updated all dependencies.
+
 ## [4.2.0]- 2024.04.03
 
 ### Changed